feat(submission): move getFilteredResponses over to submission util
